pyproject.toml
[project]
name = "kb-compiler"
version = "0.1.0"
description = "LLM Knowledge Compiler - Transform raw documents into structured wiki"
authors = [{ name = "User", email = "user@example.com" }]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "markitdown>=0.0.1a3",
    "python-frontmatter>=1.0.0",
    "watchdog>=3.0.0",
    "aiohttp>=3.9.0",
]
requires-python = ">=3.10"

[project.scripts]
kb-compiler = "kb_compiler.main:app"
qmd = "kb_compiler.qmd_cli:app"

[project.optional-dependencies]
kimi-code = [
    "anthropic>=0.20.0",
]
qmd = [
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=2.5.0",
    "flashrank>=0.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "anthropic>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
